fix(web): MessageCard を Client Component に分離 — Server Component で onClick を使えない問題を修正
Server Component の page.tsx 内に onClick ハンドラが含まれていたため
「Event handlers cannot be passed to Client Component props」エラーが
本番環境で発生していた。

MessageCard とその依存（ConfidenceMeter・ヘルパー関数・設定定数）を
message-card.tsx に切り出し "use client" を付与することで修正。
